+++ tests/_illuminate.php
@@ -116,9 +116,10 @@ abstract class IlluminateMock extends PHPUnit_Framework_TestCase
     $url->shouldReceive('to')->andReturnUsing(function ($url) {
       return $url == '#' ? $url : 'https://test/en/'.$url;
     });
-    $url->shouldReceive('route')->with('user.edit', array(2))->andReturnUsing(function ($name, array $params) {
-      return sprintf('/users/%d/edit', array_shift($params));
-    });
+    $url->shouldReceive('action')->with('UsersController@edit', array(2))->andReturn('/users/2/edit');
+    $url->shouldReceive('action')->with('UsersController@edit', 2)->andReturn('/users/2/edit');
+    $url->shouldReceive('route')->with('user.edit', array(2))->andReturn('/users/2/edit');
+    $url->shouldReceive('route')->with('user.edit', 2)->andReturn('/users/2/edit');
 
     return $url;
   }